// project_phase3_sla_analytics/frontend/src/lib/api/forms.ts
export interface FormField {
  id: string;
  type: "text" | "number" | "select" | "radio" | "checkbox" | "textarea" | "table" |"jalali_date";
  label: string;
  description?: string;
  required: boolean;
  options?: string[];
  columns?: TableColumn[]; 
}

export interface FormSchema {
  fields: FormField[];
}

export interface Form {
  id: string;
  name: string;
  customId?: string;
  description?: string;
  schema: FormSchema;
  version: number;
  isActive: boolean;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  _count?: { submissions: number };
}